修改未注册时的图标显示 增加功能：在启动界面之前检测CAD和系统是否运行，在关闭之前检查是否关闭CAD

+++ GESESApp/Registry.h
@@ -116,10 +116,10 @@ class CRegistry {
 public:
 	
 	CRegistry	(DWORD flags = CREG_CREATE);	
-	virtual		~CRegistry() { Close(); for (int i=0; i < _reEntries.size(); ++i) delete _reEntries[i]; delete [] _lpszSubKey; }
+	virtual		~CRegistry() { Close(); for (int i=0; i < (int)_reEntries.size(); ++i) delete _reEntries[i]; delete [] _lpszSubKey; }
 
 	CRegEntry&	operator[](LPCTSTR lpszVName);
-	CRegEntry*	GetAt(size_t n) { assert(n < Count());  return _reEntries.at(n); }
+	CRegEntry*	GetAt(int n) { assert(n < Count());  return _reEntries.at(n); }
 	
 	bool		Open(LPCTSTR lpszRegPath, HKEY hRootKey = HKEY_LOCAL_MACHINE,
 				DWORD dwAccess = KEY_QUERY_VALUE | KEY_SET_VALUE, bool bAuto = false);
